Create main.cpp
This game allows two players on the same computer to take turns making moves on a chess board. The game checks if each move is valid based on the rules of chess and the current state of the board. The game ends when a player captures the opponent's king or a draw is reached.
If you receive the error "string subscript out of range" when running this code, it means that the input move string is not in the correct format. The move should be entered as a string in the format "A1 B2" where "A1" is the starting position and "B2" is the ending position. The letters represent the columns (A-H) and the numbers represent the rows (1-8). Make sure to include a space between the starting and ending positions.